+++ modules/actions/exec.js
@@ -1,11 +1,11 @@
 const path = require('path');
 
-const { CONTAINER_TARGET_PATH } = require('../constants');
+const { CONTAINER_TARGET_PATH, DOCKER_SERVICE_NAME } = require('../constants');
 const { log, dockerExec, toPosixPath } = require('../helpers');
 
 const { start } = require('./start');
 const { dockerHostUserParams } = require('./docker-utils');
-const { installDependencies, mustHaveProject } = require('./utils');
+const { mustHaveProject } = require('./utils');
 
 function dockerRelativeWorkdir(libdragonInfo) {
   return (
@@ -30,7 +30,6 @@ const exec = async (libdragonInfo, commandAndParams) => {
   );
 
   const tryCmd = (libdragonInfo) =>
-    libdragonInfo.containerId &&
     dockerExec(
       libdragonInfo,
       [
@@ -47,30 +46,10 @@ const exec = async (libdragonInfo, commandAndParams) => {
     if (!started) {
       const newId = await start(libdragonInfo);
       started = true;
-
-      // Re-install vendors on new container if one was created upon start
-      // Ideally we would want the consumer to handle dependencies and rebuild
-      // libdragon if necessary. Currently this saves the day with a little bit
-      // extra waiting when the container is deleted.
-      if (libdragonInfo.containerId !== newId) {
-        await installDependencies({
-          ...libdragonInfo,
-          containerId: newId,
-        });
-      }
-      await tryCmd({
-        ...libdragonInfo,
-        containerId: newId,
-      });
+      await tryCmd(libdragonInfo);
       return newId;
     }
   };
-
-  if (!libdragonInfo.containerId) {
-    log(`Container does not exist for sure, restart`, true);
-    await startOnceAndCmd();
-    return libdragonInfo;
-  }
 
   try {
     await tryCmd(libdragonInfo);
@@ -78,7 +57,7 @@ const exec = async (libdragonInfo, commandAndParams) => {
     if (
       !e.out ||
       // TODO: is there a better way?
-      !e.out.toString().includes(libdragonInfo.containerId)
+      !e.out.toString().includes(DOCKER_SERVICE_NAME)
     ) {
       throw e;
     }