+++ modules/constants.js
@@ -1,11 +1,13 @@
 module.exports = {
   DOCKER_HUB_IMAGE: 'ghcr.io/dragonminded/libdragon:latest',
+  DOCKER_SERVICE_NAME: 'toolchain',
   LIBDRAGON_GIT: 'https://github.com/DragonMinded/libdragon',
   LIBDRAGON_BRANCH: 'trunk',
   LIBDRAGON_SUBMODULE: 'libdragon',
   CONTAINER_TARGET_PATH: '/libdragon',
   LIBDRAGON_PROJECT_MANIFEST: '.libdragon',
-  CACHED_CONTAINER_FILE: 'libdragon-docker-container',
+  LIBDRAGON_COMPOSE_FILE: 'docker-compose.json',
+
   CONFIG_FILE: 'config.json',
   DEFAULT_STRATEGY: 'submodule',
 
@@ -16,4 +18,5 @@ module.exports = {
   STATUS_VALIDATION_ERROR: 4,
 
   IMAGE_FILE: 'docker-image', // deprecated
+  CACHED_CONTAINER_FILE: 'libdragon-docker-container', // deprecated
 };